requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
apscheduler==3.11.0
google-generativeai==0.8.3
requests==2.32.3
python-dotenv==1.0.1
pydantic-settings==2.6.1
jinja2==3.1.4
moviepy==1.0.3
google-api-python-client==2.128.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
edge-tts==6.1.10
anyio==4.2.0
